+++ ocf_data_sampler/select/__init__.py
@@ -1,7 +1,6 @@
-from .fill_time_periods import fill_time_periods
-from .find_contiguous_time_periods import (
+from .time_periods import (
+    fill_time_periods,
     find_contiguous_t0_periods,
+    find_contiguous_t0_periods_nwp,
     intersection_of_multiple_dataframes_of_periods,
 )
-from .location import Location
-from ..common.indexing import get_indices_in_sorted_unique
