+++ ocf_data_sampler/select/time_slice.py
@@ -3,7 +3,7 @@
 import numpy as np
 import xarray as xr
 
-from ocf_data_sampler.select import get_indices_in_sorted_unique
+from ocf_data_sampler.common.indexing import get_indices_in_sorted_unique
 from ocf_data_sampler.common.time_utils import date_range, datetime_ceil
