+++ tests/common/test_lightarray.py
