+++ tests/select/test_spatial_slice.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from ocf_data_samplefrom ocf_data_sampler.spatial import Location
+from ocf_data_sampler.spatial import Location
 from ocf_data_sampler.select.spatial_slice import (
     _get_pixel_index_location,
     select_spatial_slice_pixels,
