+++ tests/select/test_time_periods.py
@@ -1,7 +1,8 @@
 import numpy as np
 import pandas as pd
 
-from ocf_data_sampler.select.find_contiguous_time_periods import (
+from ocf_data_sampler.select.time_periods import (
+    fill_time_periods,
     find_contiguous_t0_periods,
     find_contiguous_t0_periods_nwp,
     intersection_of_2_dataframes_of_periods,
@@ -205,3 +206,45 @@ def test_intersection_of_multiple_dataframes_of_periods():
 
     # Check if results are as expected
     assert result.equals(expected_result)
+
+
+
+def test_fill_time_periods():
+    time_periods = pd.DataFrame(
+        {
+            "start_dt": np.array([
+                "2021-01-01 04:10:00",
+                "2021-01-01 09:00:00",
+                "2021-01-01 09:15:00",
+                "2021-01-01 12:00:00",
+            ], dtype="datetime64[ns]"),
+            "end_dt": np.array([
+                "2021-01-01 06:00:00",
+                "2021-01-01 09:00:00",
+                "2021-01-01 09:20:00",
+                "2021-01-01 14:45:00",
+            ], dtype="datetime64[ns]"),
+        },
+    )
+
+    filled = fill_time_periods(time_periods, freq=np.timedelta64(30, "m"))
+
+    expected = np.array(
+        [
+            "2021-01-01 04:30",
+            "2021-01-01 05:00",
+            "2021-01-01 05:30",
+            "2021-01-01 06:00",
+            "2021-01-01 09:00",
+            "2021-01-01 12:00",
+            "2021-01-01 12:30",
+            "2021-01-01 13:00",
+            "2021-01-01 13:30",
+            "2021-01-01 14:00",
+            "2021-01-01 14:30",
+        ],
+        dtype="datetime64[ns]",
+    )
+
+    assert np.array_equal(filled, expected)
+
