+++ tests/select/test_time_slice.py
