+++ tests/test_utils.py
@@ -2,7 +2,7 @@ import dask.array
 import numpy as np
 import xarray as xr
 
-from ocf_data_sampler.load.open_xarray_tensorstore import open_zarr
+from ocf_data_sampler.common.tensorstore import open_zarr
 from ocf_data_sampler.utils import load, load_data_dict
